requirements.txt
openai
requests
pandas
openpyxl
newspaper3k
lxml_html_clean
